+++ database/migrations/2021_06_01_142440_create_imparte_table.php
@@ -16,10 +16,10 @@ class CreateImparteTable extends Migration
         Schema::create('imparte', function (Blueprint $table) {
             $table->increments('id');
              //FK
-            $table->integer('docente_id')->unsigned();
+            $table->bigInteger('docente_id')->unsigned();
             $table->foreign('docente_id')->references('id')->on('docentes');
             //FK
-            $table->integer('materia_id')->unsigned();
+            $table->bigInteger('materia_id')->unsigned();
             $table->foreign('materia_id')->references('id')->on('materias');
             $table->timestamps();
         });
