+++ database/migrations/2021_06_01_142614_create_asignadas_table.php
@@ -16,10 +16,10 @@ class CreateAsignadasTable extends Migration
         Schema::create('asignadas', function (Blueprint $table) {
             $table->increments('id');
             //FK
-            $table->integer('grupo_id')->unsigned();
+            $table->bigInteger('grupo_id')->unsigned();
             $table->foreign('grupo_id')->references('id')->on('grupos');
             //FK
-            $table->integer('materia_id')->unsigned();
+            $table->bigInteger('materia_id')->unsigned();
             $table->foreign('materia_id')->references('id')->on('materias');        
             $table->timestamps();
         });
