+++ database/migrations/2021_06_01_145036_create_otras_actividades_table.php
@@ -15,7 +15,7 @@ class CreateOtrasActividadesTable extends Migration
     {
         Schema::create('otras_actividades', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('entrevista_id')->unsigned();//alumnos
+            $table->bigInteger('entrevista_id')->unsigned();//alumnos
             $table->foreign('entrevista_id')->references('id')->on('entrevista_fresca_alumnos');
             $table->string('respuesta1');
             $table->string('respuesta2');
