+++ database/migrations/2021_06_01_161340_create_encontrar_estilo_aprendizaje_table.php
@@ -15,7 +15,7 @@ class CreateEncontrarEstiloAprendizajeTable extends Migration
     {
         Schema::create('encontrar_estilo_aprendizaje', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('test_id')->unsigned();  //test del alumno
+            $table->bigInteger('test_id')->unsigned();  //test del alumno
             $table->foreign('test_id')->references('id')->on('test');
             $table->char('respuesta1');
             $table->char('respuesta2');
