+++ database/migrations/2021_06_01_161817_create_planificacion_table.php
@@ -15,7 +15,7 @@ class CreatePlanificacionTable extends Migration
     {
         Schema::create('planificacion', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('habito_id')->unsigned();
+            $table->bigInteger('habito_id')->unsigned();
             $table->foreign('habito_id')->references('id')->on('test_habito_estudio');
             $table->char('respuesta1');
             $table->char('respuesta2');
