+++ database/migrations/2021_06_01_162000_create_estrategias_aprendizaje_table.php
@@ -15,7 +15,7 @@ class CreateEstrategiasAprendizajeTable extends Migration
     {
         Schema::create('estrategias_aprendizaje', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('habito_id')->unsigned();
+            $table->bigInteger('habito_id')->unsigned();
             $table->foreign('habito_id')->references('id')->on('test_habito_estudio');
             $table->char('respuesta1');
             $table->char('respuesta2');
