+++ database/migrations/2021_06_01_195527_create_seguimiento_alumnos_riesgo_table.php
@@ -15,9 +15,9 @@ class CreateSeguimientoAlumnosRiesgoTable extends Migration
     {
         Schema::create('seguimiento_alumnos_riesgo', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('tutor_id')->unsigned()->nullable();
+            $table->bigInteger('tutor_id')->unsigned()->nullable();
             $table->foreign('tutor_id')->references('id')->on('docentes');
-            $table->integer('alumno_id')->unsigned()->nullable();
+            $table->bigInteger('alumno_id')->unsigned()->nullable();
             $table->foreign('alumno_id')->references('nia')->on('alumnos');
             $table->date('fecha'); 
             $table->string('promedio_acumulado');
