+++ database/migrations/2021_06_01_201334_create_reporte_tutorias_table.php
@@ -15,7 +15,7 @@ class CreateReporteTutoriasTable extends Migration
     {
         Schema::create('reporte_tutorias', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('tutor_id')->unsigned()->nullable();
+            $table->bigInteger('tutor_id')->unsigned()->nullable();
             $table->foreign('tutor_id')->references('id')->on('docentes');
             ///GRUPO->MATERIA->DOCENTE
             $table->string('ciclo_escolar');
